make the hashing test local and add ij runconfig
